Fix up the windows build.
Former-commit-id: f9200e72ba700241dc81384ab7fdde5f7d00b810

+++ libdevcore/FixedHash.h
@@ -152,7 +152,7 @@ public:
 	{
 		FixedHash ret;
 		for (auto& i: ret.m_data)
-			i = std::uniform_int_distribution<uint8_t>(0, 255)(_eng);
+			i = (uint8_t)std::uniform_int_distribution<uint16_t>(0, 255)(_eng);
 		return ret;
 	}
